refactor(decomp): hoist raptor/common cycle hub to proximadb-raptor-common crate (link 8 enabler)
Move `src/storage/engines/raptor/common.rs` (1,938 LOC — the raptor cycle hub,
holding the shared types `Predicate`, `RowGroup`, `BoundaryDetectionConfig`,
`SpilloverDetectionConfig`, etc.) into a new `proximadb-raptor-common` crate.
Root `common.rs` becomes a one-line re-export shim (`pub use proximadb_raptor_common::*;`),
so every `crate::storage::engines::raptor::common::*` path — the 2 direct
importers + the `Predicate` re-export in `mod.rs` (used by ~40 files) — resolves
unchanged.

This is the **raptor cycle-hub inversion** — the single biggest unblocker for the
raptor engine extraction. With common in a crate, the raptor engine leaves
(`consolidated_reader`, `writer`, `engine.rs`) lose their `super::common` root
dependency and can eventually move to crates depending on `proximadb-raptor-common`
downward. It was gated on #952 (which cleared common's `crate::compute` quant dep
by deleting two dead `get_distances_batch_optimized` methods).

Mechanical move (history-preserving `git mv`) + import free-swaps
(`crate::proto`→`proximadb_proto`, `crate::core::bloom`→`proximadb_bloom`,
`proximacodec::types::ProximaScheme`→`proximadb_codec`). Verified hoist-ready
pre-move: no sibling deps (no `super::config`/`constants`); all deps free/crate.

Impact: ~1,938 LOC out of the root single-compile unit (the build-OOM / jobs=2
lever). Storage up-edge ratchet 181 (held — common's compute up-edges were already
cleared by #952; this move removes common's remaining core/proto refs from root
storage).

Verified: root --lib + new crate + clippy --lib --bins + cargo check --tests all
clean (tests resolve through the shim); fmt clean; layering + workspace-boundaries
pass; work-commit-check green.
