+++ Cargo.toml
@@ -78,6 +78,7 @@ members = [
     "crates/modalities/proximadb-rank-expr",
     "crates/modalities/proximadb-rank-profile",
     "crates/modalities/proximadb-rank-onnx",
+    "crates/storage/proximadb-raptor-common",
     "crates/storage/proximadb-storage-common",
     "crates/storage/proximadb-storage-ports",
     "crates/storage/proximadb-storage-traits",
@@ -599,6 +600,7 @@ jni = { version = "0.21", optional = true }
 # Node.js bindings via NAPI-RS
 napi = { version = "2.14", features = ["napi9"], optional = true }
 napi-derive = { version = "2.14", optional = true }
+proximadb-raptor-common = { version = "0.2.0", path = "crates/storage/proximadb-raptor-common" }
 
 # GPU acceleration dependencies (platform-specific)
 [target.'cfg(all(target_os = "macos", target_arch = "aarch64"))'.dependencies]
