+++ crates/storage/proximadb-raptor-common/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "proximadb-raptor-common"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Shared RAPTOR-engine types (Predicate, RowGroup, centroid/spillover detection) — hoisted out of the root crate as the raptor cycle-hub inversion (unblocks raptor engine extraction)"
+
+[lib]
+name = "proximadb_raptor_common"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+serde.workspace = true
+proximadb-proto.workspace = true
+proximadb-bloom.workspace = true
+proximadb-codec.workspace = true
+proximadb-compression.workspace = true
+proximadb-hardware-caps.workspace = true
+
+[lints]
+workspace = true
