+++ src/storage/engines/raptor/common.rs
@@ -1,1938 +1,9 @@
-use crate::proto::proximadb_v1::VectorRecord;
-use anyhow::Result;
-use proximadb_compression::CompressionAlgorithm;
-/// Common types and structures shared across RAPTOR modules
-/// This eliminates duplication between reader, writer, compaction, and other modules
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-// Re-export ProximaScheme for use in RAPTOR modules
-pub use crate::storage::engines::core::ops::proximacodec::types::ProximaScheme;
-
-// ====== Core RowGroup Structure (unified from rowgroup.rs and compaction.rs) ======
-
-/// Primary RowGroup structure used throughout RAPTOR
-/// Unified RowGroup structure consolidating all variants
-///
-/// This structure replaces:
-/// - RowGroup (common.rs) - original definition
-/// - RaptorRowGroupMetadata (common.rs) - compact representation
-/// - HybridRowGroup (rowgroup_manager.rs) - columnar variant
-/// - RowGroups duplicates (rowgroup.rs, rowgroup_manager.rs)
-///
-/// Design decisions:
-/// - Use u16 for id/centroid_id (supports 65,536 rowgroups = 67M+ vectors)
-/// - String for vector_id (supports UUIDs and custom IDs)
-/// - Removed GraphNode structures (obsolete with Matrix Trinity)
-#[derive(Debug, Clone)]
-pub struct RowGroup {
-    // Core identifiers (u16 for 67M vectors per file)
-    pub id: u16,                // Rowgroup ID == Centroid ID (1:1 mapping)
-    pub offset: u64,            // File offset for this rowgroup
-    pub compressed_size: u64,   // Compressed size in bytes
-    pub uncompressed_size: u64, // Original size
-    pub vector_count: usize,    // Number of vectors in this rowgroup
-    pub max_vectors: usize,     // Maximum capacity (from smart sizing)
-
-    // Column pages with individual compression (from RaptorRowGroupMetadata)
-    pub column_pages: HashMap<ColumnType, ColumnPageMetadata>,
-
-    // Statistics
-    pub vector_stats: VectorStats,
-    pub metadata_stats: HashMap<String, RaptorColumnStats>,
-
-    // Bloom filter for this row group's vector IDs
-    pub bloom_filter: Option<RowGroupBloomFilter>,
-    pub bloom_filter_offset: Option<u64>,
-
-    // Compression and temporal info
-    pub compression_codec: String,
-    pub min_timestamp: Option<i64>,
-    pub max_timestamp: Option<i64>,
-
-    // Centroid and clustering info (from HybridRowGroup)
-    pub centroid: Option<Vec<f32>>,            // Centroid vector
-    pub centroid_stats: Option<CentroidStats>, // Statistics for search
-
-    // Columnar storage (from HybridRowGroup)
-    // Columnar data is not serialized - it's loaded at runtime
-    pub columnar_data: Option<ColumnarBlock>, // For active/cached rowgroups
-
-    // Matrix storage (Matrix Trinity architecture)
-    pub p2_matrix: Option<P2Matrix>, // Intra-rowgroup distances
-    pub pxk_matrix: Option<VectorCentroidMatrix>, // Vector-to-centroid distances
-
-    // Cached data (for compaction/search)
-    pub vectors: Option<Vec<VectorRecord>>, // Raw vectors when loaded
-}
-
-impl RowGroup {
-    pub fn new(id: u16) -> Self {
-        Self::with_capacity(id, 1024) // Default capacity
-    }
-
-    pub fn with_capacity(id: u16, max_vectors: usize) -> Self {
-        Self {
-            id,
-            offset: 0,
-            compressed_size: 0,
-            uncompressed_size: 0,
-            vector_count: 0, // Initialize as 0
-            max_vectors,
-            column_pages: HashMap::new(),
-            vector_stats: VectorStats::default(),
-            metadata_stats: HashMap::new(),
-            bloom_filter: None,
-            bloom_filter_offset: None,
-            compression_codec: "zstd".to_string(),
-            min_timestamp: None,
-            max_timestamp: None,
-            centroid: None,
-            centroid_stats: None,
-            columnar_data: None,
-            p2_matrix: None,
-            pxk_matrix: None,
-            vectors: None,
-        }
-    }
-
-    /// Check if rowgroup has capacity for more vectors
-    pub fn has_capacity(&self) -> bool {
-        self.vector_count < self.max_vectors
-    }
-
-    /// Get available capacity
-    pub fn available_capacity(&self) -> usize {
-        self.max_vectors.saturating_sub(self.vector_count)
-    }
-
-    /// Convert to compact metadata for serialization (backward compatibility)
-    pub fn to_metadata(&self) -> RaptorRowGroupMetadata {
-        RaptorRowGroupMetadata {
-            id: self.id,
-            vector_count: self.vector_count,
-            offset: self.offset,
-            compressed_size: self.compressed_size,
-            column_pages: self.column_pages.clone(),
-            vector_stats: self.vector_stats.clone(),
-            metadata_stats: self.metadata_stats.clone(),
-            min_timestamp: self.min_timestamp,
-            max_timestamp: self.max_timestamp,
-            centroid: self.centroid.clone(),
-            centroid_stats: self.centroid_stats.clone(),
-            bloom_filter_offset: self.bloom_filter_offset,
-        }
-    }
-
-    /// Create from metadata (for deserialization)
-    pub fn from_metadata(metadata: RaptorRowGroupMetadata) -> Self {
-        Self {
-            id: metadata.id,
-            offset: 0,                           // Will be set during file reading
-            compressed_size: 0,                  // Will be calculated
-            uncompressed_size: 0,                // Will be calculated
-            vector_count: metadata.vector_count, // Using vector_count from metadata
-            max_vectors: metadata.vector_count,  // Already full
-            column_pages: metadata.column_pages,
-            vector_stats: metadata.vector_stats,
-            metadata_stats: metadata.metadata_stats,
-            bloom_filter: None, // Will be loaded on demand
-            bloom_filter_offset: None,
-            compression_codec: "zstd".to_string(),
-            min_timestamp: metadata.min_timestamp,
-            max_timestamp: metadata.max_timestamp,
-            centroid: metadata.centroid,
-            centroid_stats: metadata.centroid_stats,
-            columnar_data: None,
-            p2_matrix: None,
-            pxk_matrix: None,
-            vectors: None,
-        }
-    }
-}
-
-/// Columnar storage block within a rowgroup (dimension-major format)
-/// Migrated from HybridRowGroup for unified architecture
-#[derive(Debug, Clone)]
-pub struct ColumnarBlock {
-    /// Vector IDs (for mapping back to original)
-    pub vector_ids: Vec<String>,
-    /// Transposed vectors - each dimension is a separate array for SIMD
-    pub transposed_vectors: Option<TransposedVectors>,
-    /// Proxima encoded data for compression
-    pub proxima_data: Option<ProximaEncodedData>,
-    /// Quantized vectors if quantization is enabled
-    pub quantized_data: Option<QuantizedColumnarData>,
-    /// Metadata for each vector
-    pub metadata_columns: MetadataColumns,
-}
-
-impl Default for ColumnarBlock {
-    fn default() -> Self {
-        Self {
-            vector_ids: Vec::new(),
-            transposed_vectors: None,
-            proxima_data: None,
-            quantized_data: None,
-            metadata_columns: MetadataColumns {
-                string_columns: HashMap::new(),
-                numeric_columns: HashMap::new(),
-                boolean_columns: HashMap::new(),
-            },
-        }
-    }
-}
-
-/// Dimension-major vector storage for SIMD operations
-#[derive(Debug, Clone)]
-pub struct TransposedVectors {
-    /// Each Vec<f32> represents one dimension across all vectors
-    /// dimensions[d] = [v0[d], v1[d], v2[d], ...] for dimension d
-    pub dimensions: Vec<Vec<f32>>,
-    /// Vector dimension
-    pub dimension: usize,
-    /// Number of vectors
-    pub vector_count: usize,
-}
-
-/// Proxima encoded columnar data
-#[derive(Debug, Clone)]
-pub struct ProximaEncodedData {
-    /// Encoded dimensions using Proxima compression
-    pub encoded_dimensions: Vec<Vec<u8>>,
-    /// Encoding scheme used for each dimension
-    pub encoding_schemes: Vec<ProximaScheme>,
-    /// Compression ratio achieved
-    pub compression_ratio: f32,
-}
-
-/// Quantized columnar data
-#[derive(Debug, Clone)]
-pub struct QuantizedColumnarData {
-    /// Binary quantized vectors (1 bit per dimension)
-    pub binary: Option<Vec<Vec<u8>>>,
-    /// INT8 quantized vectors (8 bits per dimension)
-    pub int8: Option<Vec<Vec<i8>>>,
-    /// PQ4 quantized vectors (4 bits per dimension)
-    pub pq4: Option<Vec<Vec<u8>>>,
-    /// PQ8 quantized vectors (8 bits per dimension)
-    pub pq8: Option<Vec<Vec<u8>>>,
-    /// Quantization parameters
-    pub quantization_params: QuantizationParams,
-}
-
-/// Quantization parameters for columnar data
-#[derive(Debug, Clone)]
-pub struct QuantizationParams {
-    pub scale: f32,
-    pub offset: f32,
-    pub codebook: Option<Vec<Vec<f32>>>, // For PQ quantization
-}
-
-/// Metadata stored in columnar format
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MetadataColumns {
-    /// String metadata columns
-    pub string_columns: HashMap<String, Vec<Option<String>>>,
-    /// Numeric metadata columns
-    pub numeric_columns: HashMap<String, Vec<Option<f64>>>,
-    /// Boolean metadata columns
-    pub boolean_columns: HashMap<String, Vec<Option<bool>>>,
-}
-
-/// Column page metadata for selective field access
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ColumnPageMetadata {
-    pub column_type: ColumnType,
-    pub offset: u64,            // File offset for this column
-    pub compressed_size: u64,   // Compressed size in bytes
-    pub uncompressed_size: u64, // Original size
-    pub compression: CompressionAlgorithm,
-    pub encoding: Option<ProximaScheme>, // None for metadata/source content
-    pub null_count: u32,
-    pub min_value: Option<Vec<u8>>, // For pruning
-    pub max_value: Option<Vec<u8>>, // For pruning
-}
-
-/// Column types in RAPTOR files
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum ColumnType {
-    VectorsFp32,
-    VectorsQuantized,
-    Ids,
-    Metadata(String), // Key name
-    SourceContent,
-    P2Matrix,
-    PxKMatrix,
-    BloomFilter,
-}
-
-/// Backwards-compat alias for [`RaptorRowGroupMetadata`].
-pub type RowGroupMetadata = RaptorRowGroupMetadata;
-
-/// Compact metadata representation for serialization
-/// This is a lightweight version of RowGroup for storage in footer
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default)]
-pub struct RaptorRowGroupMetadata {
-    pub id: u16,             // Rowgroup ID == Centroid ID
-    pub vector_count: usize, // Number of vectors in this rowgroup
-
-    // File location information
-    pub offset: u64,          // Offset of this rowgroup in the file
-    pub compressed_size: u64, // Compressed size of this rowgroup
-
-    // Column pages with individual compression
-    pub column_pages: HashMap<ColumnType, ColumnPageMetadata>,
-
-    // Statistics
-    pub vector_stats: VectorStats,
-    pub metadata_stats: HashMap<String, RaptorColumnStats>,
-
-    // Timestamps
-    pub min_timestamp: Option<i64>,
-    pub max_timestamp: Option<i64>,
-
-    // Centroid information for fast pruning
-    pub centroid: Option<Vec<f32>>,            // Centroid vector
-    pub centroid_stats: Option<CentroidStats>, // Statistics for search optimization
-
-    // Bloom filter location (stored separately from main data)
-    pub bloom_filter_offset: Option<u64>, // File offset of bloom filter for this rowgroup
-}
-
-/// Centroid statistics for rowgroup-level pruning
-///
-/// DESIGN DECISION: We store pre-computed distances because:
-/// 1. Distance calculation for 1536-dim vectors takes ~3-5μs per vector
-/// 2. For 1000 rowgroups, that's 3-5ms just for distance calculations
-/// 3. Storage cost is minimal: ~100 bytes per rowgroup (100KB for 1000 rowgroups)
-/// 4. These stats enable O(1) pruning decisions without loading vectors
-///
-/// The stats are computed during flush/compaction when we already have vectors in memory
-///
-/// ENHANCED: Now includes all statistics needed for 5-component boosting formula
-/// These statistics are STORED at write-time and used for runtime boosting calculations
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CentroidStats {
-    pub cluster_id: u32,    // IVF cluster assignment
-    pub mean_distance: f32, // Mean distance of vectors to centroid (d₂ component)
-    pub std_deviation: f32, // Standard deviation for confidence bounds
-    pub radius: f32,        // 95th percentile distance (pruning radius)
-    pub min_distance: f32,  // Closest vector to centroid
-    pub max_distance: f32,  // Farthest vector from centroid
-
-    // === NEW: 5-Component Boosting Statistics ===
-    // Component 3: Locality/Compactness metrics
-    pub cluster_density: f32, // Vectors per unit volume (d₃ component)
-    pub intra_cluster_variance: f32, // Variance within cluster
-
-    // Component 4: Minimum inter-centroid distances
-    pub nearest_centroid_ids: Vec<u16>, // Top 5 nearest centroids by ID
-    pub nearest_centroid_distances: Vec<f32>, // Distances to nearest centroids (d₄ component)
-
-    // Component 5: Maximum inter-centroid distances
-    pub farthest_centroid_ids: Vec<u16>, // Top 5 farthest centroids by ID
-    pub farthest_centroid_distances: Vec<f32>, // Distances to farthest centroids (d₅ component)
-
-    // Boundary detection helpers (for runtime Phase 1 & 2)
-    pub boundary_vector_count: usize, // Count of vectors near cluster edge
-    pub boundary_vector_ratio: f32,   // Percentage of boundary vectors (e.g., 0.18 = 18%)
-    pub avg_spillover_ratio: f32,     // Avg d(V,C_other)/d(V,C_own) for boundary vectors
-
-    // Pre-computed bounds for common distance metrics
-    // These enable triangle inequality based pruning
-    pub euclidean_bounds: Option<DistanceBounds>,
-    pub cosine_bounds: Option<DistanceBounds>,
-    pub dot_product_bounds: Option<DistanceBounds>,
-
-    // Nearest neighbor rowgroups for multi-probe search
-    // Store top-K nearest rowgroups by centroid distance
-    // This replaces the separate CentroidNeighbors structure
-    pub neighbor_rowgroups: Vec<RowGroupNeighbor>,
-}
-
-/// Neighbor rowgroup reference for multi-probe search
-/// Lightweight reference stored in each rowgroup's metadata
-/// Only stores INDICES, not distances - distances computed at query time
-/// Optimized with u16 IDs: supports 65,536 rowgroups = 67M vectors per file
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct RowGroupNeighbor {
-    pub rowgroup_id: u16, // Direct index into footer centroids array (67M vectors max)
-    pub neighbor_cluster_id: u16, // Cluster assignment of neighbor (65k clusters max)
-    pub neighbor_type: NeighborType, // Hierarchical classification
-
-                          // NO DISTANCE STORAGE! Distances are computed at query time because:
-                          // 1. With p≥1024, intra-rowgroup connectivity is strong
-                          // 2. We use hierarchical navigation: sqrt(k) super-clusters
-                          // 3. Query-specific distances needed for accurate ranking
-                          // 4. Allows dynamic distance metrics without rewriting files
-
-                          // OPTIMAL HIERARCHICAL STRATEGY:
-                          // - Ultra-small (k≤25): All neighbors (direct access)
-                          // - Small (k≤100): Max 8 neighbors (limited exploration)
-                          // - Medium (k≤1000): 0.8×√k intra neighbors (sqrt-based)
-                          // - Large (k≤5000): 0.5×√k intra + 6 inter (balanced hierarchy)
-                          // - XLarge (k>5000): 2×ln(k) intra + 0.8×ln(k) inter (log scaling)
-                          // - Memory: 8 bytes per neighbor (33% savings vs u32)
-                          // - Performance: <100μs latency for collections up to 5M vectors
-}
-
-/// Type of neighbor relationship for hierarchical navigation
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub enum NeighborType {
-    IntraSuperCluster, // Within same super-cluster (local)
-    InterSuperCluster, // Different super-cluster (global)
-    Direct,            // For small collections (k < 100)
-}
-
-/// Pre-computed distance bounds for fast pruning
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct DistanceBounds {
-    pub min: f32, // Minimum possible distance to any vector in rowgroup
-    pub max: f32, // Maximum possible distance to any vector in rowgroup
-    pub p50: f32, // Median distance (for ranking)
-    pub p90: f32, // 90th percentile (for adaptive pruning)
-}
-
-/// Row page metadata for detailed page-level tracking
-#[derive(Debug, Clone)]
-pub struct RowPageMetadata {
-    pub page_id: u32,
-    pub file_offset: i64,
-    pub compressed_size: i64,
-    pub uncompressed_size: i64,
-    pub num_rows: i32,
-    pub first_id: Vec<u8>,
-    pub last_id: Vec<u8>,
-    pub compression_codec: String,
-}
-
-// HNSW removed - using Matrix Trinity architecture instead
-// P² matrix: Intra-rowgroup distances stored per rowgroup
-// K² matrix: Inter-centroid distances stored in footer
-// P×K matrix: Vector-to-centroid distances stored per rowgroup
-
-// ====== Vector Statistics (unified) ======
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct VectorStats {
-    pub dimension: usize,
-    pub min_norm: f32,
-    pub max_norm: f32,
-    pub centroid: Vec<f32>,
-    pub quantization_error: Option<f32>,
-    pub encoding: VectorEncoding,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default)]
-pub enum VectorEncoding {
-    #[default]
-    Raw,
-    ProductQuantization {
-        num_subvectors: usize,
-        bits_per_subvector: usize,
-    },
-    ScalarQuantization {
-        bits: usize,
-        scale: f32,
-        zero_point: f32,
-    },
-    Binary,
-    Proxima {
-        scheme: ProximaScheme,
-    },
-}
-
-// ====== Column Statistics (unified from reader.rs and rowgroup.rs) ======
-
-/// Backwards-compat alias for [`RaptorColumnStats`].
-pub type ColumnStats = RaptorColumnStats;
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RaptorColumnStats {
-    pub null_count: usize,
-    pub distinct_count: Option<usize>,
-    pub min_value: Option<MetadataValue>,
-    pub max_value: Option<MetadataValue>,
-    pub encoding: ColumnEncoding,
-    pub compressed_size: usize,
-    pub uncompressed_size: usize,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub enum ColumnEncoding {
-    Dictionary { num_entries: usize },
-    Integer { bits: usize },
-    Float,
-    Boolean,
-    String,
-    Proxima { scheme: ProximaScheme },
-}
-
-// ====== Metadata Column (unified from reader.rs and writer.rs) ======
-
-#[derive(Debug, Clone)]
-pub struct MetadataColumn {
-    pub name: String,
-    pub data_type: MetadataData,
-    pub encoding: ColumnEncoding,
-    pub stats: RaptorColumnStats,
-    pub dictionary: Option<Vec<String>>, // For dictionary encoding
-    pub offset: u64,
-    pub size: u64,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub enum MetadataData {
-    Boolean,
-    Integer,
-    Float,
-    String,
-    List(Box<MetadataData>),
-    Map(Box<MetadataData>, Box<MetadataData>),
-}
-
-// Use proto-generated SqlValue instead of duplicate MetadataValue
-pub use crate::proto::proximadb_v1::SqlValue as MetadataValue;
-
-// PartialOrd for SqlValue is handled by proto-generated implementation
-
-// ====== Proxima Encoding Schemes (shared) ======
-
-// ProximaScheme moved to crate::storage::engines::core::ops::proximaencoder
-// Use that unified implementation instead of this duplicate
-
-// ====== Compaction Configuration (moved from config.rs duplicate) ======
-// Note: Using the one from config.rs as the source of truth
-
-// ====== File Metadata (unified) ======
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-/// CONSOLIDATED RaptorFileMetadata - Single source of truth
-/// This combines all fields from the three duplicate definitions
-pub struct RaptorFileMetadata {
-    // Core file metadata
-    pub version: u32,
-    pub created_at: i64,
-    pub created_by: String,
-    pub file_path: String,
-    pub file_size: u64,
-
-    // Row and vector counts
-    pub total_rows: usize,
-    pub total_vectors: usize,
-    pub dimension: usize,
-
-    // Collection info
-    pub collection_id: String,
-
-    // Row groups
-    pub row_groups: Vec<RaptorRowGroupMetadata>,
-    pub num_rowgroups: usize,
-    pub rowgroup_offsets: Vec<u64>,
-    pub rowgroup_sizes: Vec<u64>,
-    pub rowgroup_vector_counts: Vec<usize>,
-
-    // Schema
-    pub schema: SchemaDescriptor,
-
-    // Bloom filter metadata (for per-rowgroup bloom filters)
-    pub bloom_filter_metadata: Option<RaptorBloomFilterMetadata>,
-
-    // Compression
-    pub compression_codec: String,
-    pub compression_ratio: f64,
-
-    // Clustering metadata
-    pub cluster_centroids: Vec<Vec<f32>>,
-    pub cluster_assignments: HashMap<String, usize>,
-
-    // Metadata storage
-    pub custom_metadata: HashMap<String, String>,
-    pub key_value_metadata: Vec<KeyValue>,
-
-    // Additional metadata fields
-    pub footer_offset: u64,
-    pub footer_size: u64,
-    pub last_accessed: i64,
-    pub locality_clusters: Vec<Vec<String>>,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct SchemaDescriptor {
-    pub vector_dimension: usize,
-    pub metadata_fields: Vec<FieldDescriptor>,
-    pub version: u32,
-}
-
-impl Default for SchemaDescriptor {
-    fn default() -> Self {
-        Self {
-            vector_dimension: 384, // Default OpenAI embedding dimension
-            metadata_fields: Vec::new(),
-            version: 1,
-        }
-    }
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct KeyValue {
-    pub key: String,
-    pub value: String,
-}
-
-#[derive(Debug, Clone)]
-pub struct LocalityClusterInfo {
-    pub cluster_id: u32,
-    pub start_offset: u64,
-    pub size: u64,
-    pub vector_count: usize,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct FieldDescriptor {
-    pub name: String,
-    pub data_type: MetadataData,
-    pub nullable: bool,
-    pub default_value: Option<MetadataValue>,
-}
-
-/// Backwards-compat alias for [`RaptorBloomFilterMetadata`].
-pub type BloomFilterMetadata = RaptorBloomFilterMetadata;
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct RaptorBloomFilterMetadata {
-    pub num_bits: usize,
-    pub num_hashes: usize,
-    pub num_entries: u32, // Number of entries in the bloom filter
-    pub false_positive_rate: f64,
-    pub offset: u64,
-    pub size: u64,
-}
-
-// ====== I/O Strategy (for unified reader) ======
-
-#[derive(Debug, Clone)]
-pub struct IoStrategy {
-    pub prefetch_size: usize,
-    pub cache_policy: CachePolicy,
-    pub read_pattern: ReadPattern,
-}
-
-#[derive(Debug, Clone)]
-pub enum CachePolicy {
-    LRU,
-    LFU,
-    Cost, // Based on I/O cost
-    None,
-}
-
-#[derive(Debug, Clone)]
-pub enum ReadPattern {
-    Sequential,
-    Random,
-    Strided { stride: usize },
-    Adaptive, // Detect pattern at runtime
-}
-
-// ====== Search Results ======
-// Note: Using OptimizedSearchRecord from core::search::results instead of custom SearchResult struct
-
-// ====== Predicates for filtering ======
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct Predicate {
-    pub field: String,
-    pub op: PredicateOp,
-    pub value: MetadataValue,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum PredicateOp {
-    Eq,
-    Ne,
-    Lt,
-    Lte,
-    Gt,
-    Gte,
-    In,
-    NotIn,
-    Contains,
-    StartsWith,
-}
-
-// ====== Bloom Filter Structures ======
-
-/// Per-RowGroup bloom filter for fast membership testing
-/// Optimized for scattered ID access patterns
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RowGroupBloomFilter {
-    /// Bloom filter bits (typically 10 bits per ID for 1% false positive)
-    pub bits: Vec<u8>,
-
-    /// Number of hash functions (typically 7 for optimal)
-    pub num_hashes: usize,
-
-    /// Number of IDs in this row group
-    pub num_ids: usize,
-
-    /// Size in bits
-    pub size_bits: usize,
-
-    /// Target false positive rate
-    pub false_positive_rate: f64,
-}
-
-impl RowGroupBloomFilter {
-    /// Create new bloom filter for VectorRecord IDs
-    /// Uses core bloom filter module with delegation pattern
-    pub fn new(expected_ids: usize, false_positive_rate: f64) -> Self {
-        use crate::core::bloom::factory::BloomFilterFactory;
-        use crate::core::bloom::{BloomFilterConfig, BloomStrategy, HashAlgorithm};
-
-        // Create config optimized for ID filtering
-        let config = BloomFilterConfig {
-            strategy: BloomStrategy::ByteAligned,
-            bits_per_key: Self::calculate_bits_per_key(false_positive_rate),
-            false_positive_rate: Some(false_positive_rate),
-            expected_items: expected_ids,
-            enabled: true,
-            hash_algorithm: HashAlgorithm::Murmur3,
-        };
-
-        // Create the underlying bloom filter
-        let filter = BloomFilterFactory::create(&config);
-
-        // Extract parameters
-        let size_bits = filter.bit_count();
-        let num_hashes = filter.hash_count();
-
-        // Serialize to get bits (temporary approach - will be populated during writes)
-        let bits = filter
-            .serialize()
-            .unwrap_or_else(|_| vec![0u8; size_bits / 8])
-            .get(8..) // Skip header bytes
-            .unwrap_or(&[])
-            .to_vec();
-
-        Self {
-            bits,
-            num_hashes,
-            num_ids: 0,
-            size_bits,
-            false_positive_rate,
-        }
-    }
-
-    /// Calculate optimal bits per key for given false positive rate
-    /// Formula: k = ln(2) * m/n where k=bits_per_key, m=total_bits, n=items
-    fn calculate_bits_per_key(false_positive_rate: f64) -> u32 {
-        let bits = (-false_positive_rate.ln() / (2.0_f64.ln().powi(2))).ceil();
-        (bits as u32).clamp(4, 32) // Clamp between 4 and 32 bits
-    }
-
-    /// Insert VectorRecord ID into bloom filter
-    /// Delegates to core bloom filter for actual bit manipulation
-    pub fn insert(&mut self, vector_id: &str) -> anyhow::Result<()> {
-        use crate::core::bloom::factory::BloomFilterFactory;
-        use crate::core::bloom::{BloomFilterConfig, BloomStrategy, HashAlgorithm};
-
-        // Recreate filter from current state (for now - optimization needed)
-        let config = BloomFilterConfig {
-            strategy: BloomStrategy::ByteAligned,
-            bits_per_key: Self::calculate_bits_per_key(self.false_positive_rate),
-            false_positive_rate: Some(self.false_positive_rate),
-            expected_items: (self.num_ids + 1000).max(100), // Account for growth
-            enabled: true,
-            hash_algorithm: HashAlgorithm::Murmur3,
-        };
-
-        let mut filter = BloomFilterFactory::create(&config);
-
-        // Deferred: Restore previous state from self.bits (optimization for later)
-
-        // Insert the new ID
-        filter.insert(vector_id.as_bytes());
-
-        // Update our state
-        self.num_ids += 1;
-        self.bits = filter.serialize()?.get(8..).unwrap_or(&[]).to_vec();
-        self.size_bits = filter.bit_count();
-        self.num_hashes = filter.hash_count();
-
-        Ok(())
-    }
-
-    /// Check if VectorRecord ID might exist in this row group
-    /// Returns true if ID might exist, false if definitely doesn't exist
-    pub fn contains(&self, vector_id: &str) -> bool {
-        use crate::core::bloom::factory::BloomFilterFactory;
-        use crate::core::bloom::{BloomFilterConfig, BloomStrategy, HashAlgorithm};
-
-        if self.bits.is_empty() || self.num_ids == 0 {
-            return false;
-        }
-
-        // Recreate filter from serialized bits
-        let config = BloomFilterConfig {
-            strategy: BloomStrategy::ByteAligned,
-            bits_per_key: Self::calculate_bits_per_key(self.false_positive_rate),
-            false_positive_rate: Some(self.false_positive_rate),
-            expected_items: self.num_ids.max(100),
-            enabled: true,
-            hash_algorithm: HashAlgorithm::Murmur3,
-        };
-
-        let _filter = BloomFilterFactory::create(&config);
-
-        // Deferred: Restore state from self.bits (optimization needed)
-        // For now, this is a simplified implementation
-
-        // Use bit manipulation directly as fallback
-        self.hash_based_contains(vector_id)
-    }
-
-    /// Fallback hash-based membership test
-    /// Uses murmur3 hash with multiple hash functions
-    fn hash_based_contains(&self, vector_id: &str) -> bool {
-        if self.bits.is_empty() {
-            return false;
-        }
-
-        let key_bytes = vector_id.as_bytes();
-        let bit_array_size = self.bits.len() * 8;
-
-        if bit_array_size == 0 {
-            return false;
-        }
-
-        // Use multiple hash functions
-        for i in 0..self.num_hashes {
-            let hash = self.murmur3_hash(key_bytes, i as u32);
-            let bit_index = (hash % bit_array_size as u32) as usize;
-
-            let byte_index = bit_index / 8;
-            let bit_offset = bit_index % 8;
-
-            if byte_index >= self.bits.len() {
-                return false;
-            }
-
-            let byte_value = self.bits[byte_index];
-            let bit_set = (byte_value >> bit_offset) & 1;
-
-            if bit_set == 0 {
-                return false; // Definitely not present
-            }
-        }
-
-        true // Might be present
-    }
-
-    /// Simple murmur3 hash implementation
-    fn murmur3_hash(&self, key: &[u8], seed: u32) -> u32 {
-        const C1: u32 = 0xcc9e2d51;
-        const C2: u32 = 0x1b873593;
-        const R1: u32 = 15;
-        const R2: u32 = 13;
-        const M: u32 = 5;
-        const N: u32 = 0xe6546b64;
-
-        let mut hash = seed;
-        let mut i = 0;
-
-        // Process 4-byte chunks
-        while i + 4 <= key.len() {
-            let k = u32::from_le_bytes([key[i], key[i + 1], key[i + 2], key[i + 3]]);
-            let k = k.wrapping_mul(C1);
-            let k = k.rotate_left(R1);
-            let k = k.wrapping_mul(C2);
-
-            hash ^= k;
-            hash = hash.rotate_left(R2);
-            hash = hash.wrapping_mul(M).wrapping_add(N);
-
-            i += 4;
-        }
-
-        // Process remaining bytes
-        if i < key.len() {
-            let mut k = 0u32;
-            for j in (i..key.len()).rev() {
-                k = (k << 8) | key[j] as u32;
-            }
-            k = k.wrapping_mul(C1);
-            k = k.rotate_left(R1);
-            k = k.wrapping_mul(C2);
-            hash ^= k;
-        }
-
-        hash ^= key.len() as u32;
-        hash ^= hash >> 16;
-        hash = hash.wrapping_mul(0x85ebca6b);
-        hash ^= hash >> 13;
-        hash = hash.wrapping_mul(0xc2b2ae35);
-        hash ^= hash >> 16;
-
-        hash
-    }
-
-    /// Get memory usage in bytes
-    pub fn memory_usage(&self) -> usize {
-        std::mem::size_of::<Self>() + self.bits.len()
-    }
-
-    /// Create bloom filter for batch of IDs
-    pub fn from_ids(ids: &[String], false_positive_rate: f64) -> anyhow::Result<Self> {
-        let mut filter = Self::new(ids.len(), false_positive_rate);
-
-        for id in ids {
-            filter.insert(id)?;
-        }
-
-        Ok(filter)
-    }
-
-    /// Get statistics for this bloom filter
-    pub fn stats(&self) -> RaptorBloomFilterStats {
-        RaptorBloomFilterStats {
-            num_ids: self.num_ids,
-            size_bytes: self.memory_usage(),
-            size_bits: self.size_bits,
-            num_hashes: self.num_hashes,
-            false_positive_rate: self.false_positive_rate,
-        }
-    }
-}
-
-/// Backwards-compat alias for [`RaptorBloomFilterStats`].
-pub type BloomFilterStats = RaptorBloomFilterStats;
-
-/// Statistics for bloom filter performance tracking
-#[derive(Debug, Clone)]
-pub struct RaptorBloomFilterStats {
-    pub num_ids: usize,
-    pub size_bytes: usize,
-    pub size_bits: usize,
-    pub num_hashes: usize,
-    pub false_positive_rate: f64,
-}
-
-/// Columnar ID index within row group for SIMD scanning
-#[derive(Debug, Clone)]
-pub struct ColumnnarIdIndex {
-    /// IDs stored columnar for SIMD scanning
-    pub ids: Vec<String>,
-
-    /// Pre-computed hashes for faster comparison (optional)
-    pub id_hashes: Option<Vec<u64>>,
-
-    /// Offsets to full row data within row group
-    pub row_offsets: Vec<u32>,
-
-    /// Whether IDs are sorted (enables binary search)
-    pub is_sorted: bool,
-}
-
-// ====== Locality clustering for spatial organization ======
-
-#[derive(Debug, Clone)]
-pub struct LocalityCluster {
-    pub id: u32,
-    pub centroid: Vec<f32>,
-    pub radius: f32,
-    pub vector_ids: Vec<String>,
-    pub rowgroup_ids: Vec<u32>,
-}
-
-// ====== Centralized Footer for Centroid Storage ======
-
-/// Centralized distance matrices storage implementing complete P² + K² + P×K design
-/// All centroids and distance matrices stored columnar-encoded for O(1) access
-///
-/// COMPLETE DESIGN FORMULA: P² + K² + P×K
-/// - P²: Vectors stored in rowgroups (existing implementation)
-/// - K²: Inter-centroid distance matrix (k×k matrix)  
-/// - P×K: Vector-to-centroid distance matrix (p×k matrix per rowgroup)
-///
-/// MEMORY COMPARISON (for k=1000 rowgroups, p=1000 vectors/rowgroup, d=1536 dimensions):
-/// - Centroids (K): 1000 × 1536 × 4 = 6MB
-/// - Inter-centroid distances (K²): 1000 × 1000 × 4 = 4MB  
-/// - Vector-centroid distances (P×K): 1000 × 1000 × 4 = 4MB per rowgroup
-/// - Total navigation overhead: 6MB + 4MB + (4MB × selective loading) = ~14MB active
-///
-/// I/O BENEFITS:
-/// - Single read loads ALL centroids and K×K matrix (one 10MB I/O)
-/// - P×K matrices loaded on-demand per active rowgroup
-/// - Cached indefinitely (file doesn't change, footer doesn't change)
-/// - OS page cache or memory-mapped for zero-copy access
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RaptorFooter {
-    /// All centroids sorted by rowgroup_id for O(1) indexing
-    /// Stored using Proxima columnar encoding for compression
-    pub centroids: ColumnarCentroids,
-
-    /// K×K inter-centroid distance matrix for O(1) cluster-to-cluster distance lookup
-    /// Essential for d₂ component in 5-component boosting formula
-    /// Size: k×k×4 bytes (4MB for k=1000, compressed ~2MB with Proxima)
-    pub inter_centroid_distances: InterCentroidMatrix,
-
-    /// P×K vector-to-centroid distance matrices per rowgroup
-    /// Essential for d₁, d₄, d₅ components in 5-component boosting formula
-    /// Stored as offsets - actual matrices loaded on-demand during search
-    pub vector_centroid_matrices: Vec<VectorCentroidMatrixRef>,
-
-    /// PERFECTED: 1-to-1 Centroid-to-Rowgroup mapping for perfect parallelism  
-    /// K centroids = K rowgroups (centroid_id == rowgroup_id)
-    /// Overflow creates new centroids for balanced distribution
-    /// This enables perfect search parallelism across vector space subdivisions
-    pub total_centroids: u32, // K centroids = K rowgroups
-
-    /// Version for backward compatibility
-    pub version: u32,
-
-    /// Checksum for integrity verification
-    pub checksum: u64,
-
-    /// File metadata (already exists, just reference it)
-    pub file_metadata: RaptorFileMetadata,
-}
-
-/// Columnar-encoded centroids using Proxima
-/// Vectors are transposed for better compression and SIMD operations
-///
-/// ENCODING STRATEGY:
-/// 1. Transpose: Convert k×d matrix to d×k (better compression per dimension)
-/// 2. Delta encode each dimension (values often similar across centroids)
-/// 3. Bit-pack based on range (many dimensions need only 8-16 bits)
-/// 4. SIMD-friendly layout for fast distance calculations
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct ColumnarCentroids {
-    /// Number of centroids (typically k < 1000)
-    pub count: u32,
-
-    /// Dimension of each centroid (typically same as vector dimension)
-    pub dimension: usize,
-
-    /// Rowgroup IDs in sorted order for O(1) access
-    /// Size: k * 2 bytes (50% savings vs u32)
-    pub rowgroup_ids: Vec<u16>,
-
-    /// Transposed centroid data for columnar encoding
-    /// Layout: [dim0_values..., dim1_values..., ...]
-    /// Size: k * d * 4 bytes (before compression)
-    pub transposed_data: Vec<f32>,
-
-    /// Proxima encoding metadata for each dimension
-    pub encoding_metadata: Vec<ProximaMetadata>,
-}
-
-impl ColumnarCentroids {
-    /// Get centroid by rowgroup_id with O(1) access
-    pub fn get_centroid(&self, rowgroup_id: u16) -> Option<Vec<f32>> {
-        // Binary search since rowgroup_ids are sorted
-        match self.rowgroup_ids.binary_search(&rowgroup_id) {
-            Ok(idx) => {
-                // Reconstruct centroid from transposed data
-                let mut centroid = Vec::with_capacity(self.dimension);
-                for dim in 0..self.dimension {
-                    let offset = dim * self.count as usize + idx;
-                    centroid.push(self.transposed_data[offset]);
-                }
-                Some(centroid)
-            }
-            Err(_) => None,
-        }
-    }
-
-    /// Decode all centroids for batch operations
-    pub fn decode_all(&self) -> Vec<(u16, Vec<f32>)> {
-        let mut centroids = Vec::with_capacity(self.count as usize);
-
-        for (idx, &rowgroup_id) in self.rowgroup_ids.iter().enumerate() {
-            let mut centroid = Vec::with_capacity(self.dimension);
-            for dim in 0..self.dimension {
-                let offset = dim * self.count as usize + idx;
-                centroid.push(self.transposed_data[offset]);
-            }
-            centroids.push((rowgroup_id, centroid));
-        }
-
-        centroids
-    }
-}
-
-/// Calculate optimal number of neighbors using performance-tested formula
-/// Returns (intra_neighbors, inter_neighbors) based on collection size k
-pub fn calculate_optimal_neighbors(k: usize) -> (usize, usize) {
-    match k {
-        // Ultra-small: direct access for maximum accuracy
-        k if k <= 25 => (k.saturating_sub(1), 0),
-
-        // Small: limited neighbors to prevent over-exploration
-        k if k <= 100 => (8.min(k.saturating_sub(1)), 0),
-
-        // Medium: sqrt-based intra-cluster only (efficient single-tier)
-        k if k <= 1000 => {
-            let intra = ((k as f64).sqrt() * 0.8).ceil() as usize;
-            (intra, 0)
-        }
-
-        // Large: balanced intra + limited inter (two-tier hierarchy)
-        k if k <= 5000 => {
-            let intra = ((k as f64).sqrt() * 0.5).ceil() as usize;
-            let inter = 6; // Fixed small number for global exploration
-            (intra, inter)
-        }
-
-        // XLarge: logarithmic scaling prevents neighbor explosion
-        _ => {
-            let intra = ((k as f64).ln() * 2.0).ceil() as usize;
-            let inter = ((k as f64).ln() * 0.8).ceil() as usize;
-            (intra, inter)
-        }
-    }
-}
-
-/// Calculate number of super-clusters for hierarchical organization
-pub fn calculate_super_clusters(k: usize) -> usize {
-    match k {
-        k if k <= 100 => 1, // No super-clustering needed for small collections
-        k if k <= 1000 => ((k as f64).sqrt() / 2.0).ceil() as usize,
-        _ => ((k as f64).sqrt() / 3.0).ceil() as usize, // More super-clusters for large collections
-    }
-}
-
-/// Predict search latency in microseconds for performance planning
-pub fn predict_search_latency(k: usize, dimension: usize) -> f64 {
-    // Centroid computation: ~0.006μs per centroid for 384d (measured)
-    let centroid_latency = (k as f64) * (dimension as f64) * 0.000015;
-
-    // Neighbor exploration: 3 candidates × neighbors × 50ns per distance
-    let (intra, inter) = calculate_optimal_neighbors(k);
-    let neighbor_latency = 3.0 * (intra + inter) as f64 * 0.05;
-
-    centroid_latency + neighbor_latency
-}
-
-/// K×K Inter-centroid distance matrix for O(1) cluster navigation
-/// Heavily optimized storage using upper triangle + quantization + Proxima
-///
-/// STORAGE OPTIMIZATIONS (4-stage compression pipeline):
-/// 1. **Upper Triangle Only**: Store only [i][j] where j > i (exactly 50% savings)
-///    - Symmetric matrix property: distance(i,j) = distance(j,i)
-///    - Diagonal elements always 0.0 (not stored)
-///    - Elements stored: k*(k-1)/2 instead of k*k
-/// 2. **16-bit Quantization**: f32 → u16 with dynamic range scaling (50% savings)
-///    - Scale factor: (max_dist - min_dist) / 65535
-///    - Accuracy loss: <0.1% for typical centroid distances
-/// 3. **Delta Encoding**: From minimum distance (additional compression)
-/// 4. **Proxima Bit-packing**: Based on actual value distribution (future)
-///
-/// FINAL SIZE CALCULATION:
-/// - Original: k×k×4 bytes = 1000×1000×4 = 4MB
-/// - Upper triangle: k×(k-1)/2×2 = 1000×999/2×2 = 999KB  
-/// - With Proxima: ~500KB (estimated 50% additional compression)
-/// - **Total compression: 4MB → 500KB (87.5% space savings)**
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct InterCentroidMatrix {
-    /// Number of centroids (k)
-    pub num_centroids: u32,
-
-    /// Compressed upper triangle matrix data
-    /// Layout: Proxima-encoded distances in row-major upper triangle order
-    pub compressed_data: Vec<u8>,
-
-    /// Compression metadata for reconstruction
-    pub compression_metadata: InterCentroidCompressionMetadata,
-
-    /// Quick lookup table for O(1) access to compressed positions
-    /// Maps (row, col) → compressed_data offset
-    pub lookup_table: Vec<u32>,
-}
-
-impl InterCentroidMatrix {
-    /// Get distance between two centroids with O(1) lookup
-    /// Returns 0.0 for diagonal elements, reconstructs from compressed upper triangle otherwise
-    ///
-    /// OPTIMIZATION: Only stores upper triangle where j > i (exactly 50% space savings)
-    /// Matrix symmetry: distance(i,j) = distance(j,i), diagonal = 0.0
-    pub fn get_distance(&self, centroid_i: usize, centroid_j: usize) -> f32 {
-        if centroid_i == centroid_j {
-            return 0.0; // Diagonal elements are always 0
-        }
-
-        // Ensure upper triangle access (swap if needed to guarantee j > i)
-        let (i, j) = if centroid_i < centroid_j {
-            (centroid_i, centroid_j)
-        } else {
-            (centroid_j, centroid_i)
-        };
-
-        // Calculate upper triangle index: sum of previous rows + position in current row
-        // For row i, we store elements [i][i+1], [i][i+2], ..., [i][k-1]
-        // Total elements before row i: i*(2k-i-1)/2
-        // Position in row i: (j - i - 1)
-        let total_before_row_i = i * (2 * self.num_centroids as usize - i - 1) / 2;
-        let position_in_row_i = j - i - 1;
-        let linear_index = total_before_row_i + position_in_row_i;
-
-        // O(1) lookup using optimized indexing (no lookup table needed!)
-        self.decompress_single_distance_at_index(linear_index)
-    }
-
-    /// Create new InterCentroidMatrix from full distance matrix
-    /// Extracts and compresses only the upper triangle for optimal storage
-    pub fn from_full_matrix(distances: &[Vec<f32>]) -> Self {
-        let k = distances.len();
-        let mut upper_triangle_data = Vec::new();
-        let mut compression_metadata = InterCentroidCompressionMetadata::default();
-
-        // Extract upper triangle in row-major order
-        let mut min_dist = f32::MAX;
-        let mut max_dist = f32::MIN;
-
-        for (i, dist_row) in distances.iter().enumerate().take(k) {
-            for &dist in dist_row.iter().take(k).skip(i + 1) {
-                // Only j > i (strict upper triangle)
-                upper_triangle_data.push(dist);
-                min_dist = min_dist.min(dist);
-                max_dist = max_dist.max(dist);
-            }
-        }
-
-        // Update compression metadata
-        compression_metadata.min_distance = min_dist;
-        compression_metadata.max_distance = max_dist;
-        compression_metadata.scale_factor = (max_dist - min_dist) / 65535.0; // 16-bit quantization
-
-        // Compress using Proxima (placeholder - actual implementation would compress)
-        let compressed_data =
-            Self::compress_upper_triangle(&upper_triangle_data, &compression_metadata);
-
-        Self {
-            num_centroids: k as u32,
-            compressed_data,
-            compression_metadata,
-            lookup_table: Vec::new(), // Not needed with optimized indexing formula
-        }
-    }
-
-    /// Calculate exact storage requirement for upper triangle
-    /// Formula: k*(k-1)/2 elements (exactly 50% of full k*k matrix)
-    pub fn upper_triangle_size(k: usize) -> usize {
-        k * (k - 1) / 2
-    }
-
-    /// Decompress single distance value at specific linear index in upper triangle
-    /// Uses 16-bit quantization reconstruction
-    /// Quantization: q = (d - min) * scale, where scale = 65535 / (max - min)
-    /// Dequantization: d = min + q / scale
-    fn decompress_single_distance_at_index(&self, linear_index: usize) -> f32 {
-        if linear_index * 2 >= self.compressed_data.len() {
-            return 0.0; // Out of bounds
-        }
-
-        // Extract 16-bit quantized value from compressed data
-        let offset = linear_index * 2; // 2 bytes per 16-bit value
-        let quantized = u16::from_le_bytes([
-            self.compressed_data[offset],
-            self.compressed_data[offset + 1],
-        ]);
-
-        // Reconstruct f32 distance from quantized value
-        // Formula: d = min + q / scale_factor (scale_factor = 65535 / range)
-        self.compression_metadata.min_distance
-            + (quantized as f32 / self.compression_metadata.scale_factor)
-    }
-
-    /// Compress upper triangle data using 16-bit quantization + optional Proxima
-    fn compress_upper_triangle(
-        data: &[f32],
-        metadata: &InterCentroidCompressionMetadata,
-    ) -> Vec<u8> {
-        let mut compressed = Vec::with_capacity(data.len() * 2); // 2 bytes per f32
-
-        for &distance in data {
-            // Quantize to 16-bit
-            let normalized = (distance - metadata.min_distance) / metadata.scale_factor;
-            let quantized = normalized.clamp(0.0, 65535.0) as u16;
-            compressed.extend(&quantized.to_le_bytes());
-        }
-
-        // Deferred: Apply Proxima bit-packing for further compression
-        // For now, return 16-bit quantized data (already 50% space savings)
-        compressed
-    }
-
-    /// Decompress entire upper triangle for batch operations
-    /// Reconstructs symmetric matrix from stored upper triangle only
-    pub fn decompress_all(&self) -> Vec<Vec<f32>> {
-        let k = self.num_centroids as usize;
-        let mut matrix = vec![vec![0.0f32; k]; k];
-
-        // Diagonal elements remain 0.0 (already initialized)
-
-        // Reconstruct from upper triangle storage
-        #[allow(clippy::needless_range_loop)]
-        for i in 0..k {
-            for j in (i + 1)..k {
-                // Only upper triangle j > i
-                let distance = self.get_distance(i, j);
-                matrix[i][j] = distance;
-                matrix[j][i] = distance; // Symmetric assignment
-            }
-        }
-
-        matrix
-    }
-
-    /// Get memory footprint in bytes for the compressed upper triangle
-    pub fn memory_footprint(&self) -> usize {
-        std::mem::size_of::<Self>()
-            + self.compressed_data.len()
-            + self.compression_metadata.memory_footprint()
-    }
-}
-
-/// Compression type for matrix quantization
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
-pub enum CompressionType {
-    /// 8-bit quantization (1 byte per value)
-    Quantized8Bit,
-    /// 16-bit quantization (2 bytes per value)
-    Quantized16Bit,
-    /// 32-bit full precision (4 bytes per value)
-    FullPrecision,
-}
-
-/// Compression metadata for inter-centroid matrix reconstruction
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct InterCentroidCompressionMetadata {
-    /// Minimum distance value (for delta encoding base)
-    pub min_distance: f32,
-
-    /// Maximum distance value (for range calculation)
-    pub max_distance: f32,
-
-    /// Quantization scale factor (16-bit → f32 reconstruction)
-    pub scale_factor: f32,
-
-    /// Compression type used for quantization
-    pub compression_type: CompressionType,
-
-    /// Proxima encoding scheme per row (may vary based on distance distribution)
-    pub row_encodings: Vec<ProximaScheme>,
-
-    /// Compressed size per row for offset calculation
-    pub row_compressed_sizes: Vec<u16>,
-}
-
-impl Default for InterCentroidCompressionMetadata {
-    fn default() -> Self {
-        Self {
-            min_distance: 0.0,
-            max_distance: 1.0,
-            scale_factor: 1.0 / 65535.0,
-            compression_type: CompressionType::Quantized16Bit,
-            row_encodings: Vec::new(),
-            row_compressed_sizes: Vec::new(),
-        }
-    }
-}
-
-impl InterCentroidCompressionMetadata {
-    /// Calculate memory footprint of compression metadata
-    pub fn memory_footprint(&self) -> usize {
-        std::mem::size_of::<Self>()
-            + self.row_encodings.len() * std::mem::size_of::<ProximaScheme>()
-            + self.row_compressed_sizes.len() * std::mem::size_of::<u16>()
-    }
-}
-
-/// P×K Vector-to-centroid distance matrix for a specific rowgroup
-/// Contains distances from all P vectors in rowgroup to all K centroids
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct VectorCentroidMatrix {
-    /// Rowgroup this matrix belongs to
-    pub rowgroup_id: u16,
-
-    /// Number of vectors in this rowgroup (P)
-    pub num_vectors: u32,
-
-    /// Number of centroids (K, same across all rowgroups)
-    pub num_centroids: u32,
-
-    /// Storage strategy used for this matrix
-    pub storage_strategy: VectorCentroidStorageStrategy,
-
-    /// Compressed matrix data
-    pub compressed_data: Vec<u8>,
-
-    /// Compression metadata
-    pub compression_metadata: VectorCentroidCompressionMetadata,
-
-    /// Hierarchical data (when using hierarchical strategy)
-    pub hierarchical_data: Option<HierarchicalData>,
-
-    /// Sparse data (when using sparse strategy)
-    pub sparse_data: Option<SparseData>,
-}
-
-/// Storage strategy for P×K matrices based on sparsity and size
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub enum VectorCentroidStorageStrategy {
-    /// Store full P×K matrix (dense, small rowgroups)
-    Full,
-
-    /// Store mean + sparse deltas (medium sparsity)
-    Hierarchical,
-
-    /// Store only top-k entries per vector (high sparsity)
-    Sparse,
-}
-
-/// Hierarchical storage data for mean + sparse deltas
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct HierarchicalData {
-    /// Mean distance per centroid (K values)
-    pub mean_distances: Vec<f32>,
-
-    /// Sparse deltas from mean (only significant deviations)
-    pub sparse_deltas: Vec<DeltaEntry>,
-}
-
-/// Delta entry for hierarchical storage
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DeltaEntry {
-    /// Vector index within rowgroup
-    pub vector_index: u32,
-
-    /// Centroid index
-    pub centroid_index: u16,
-
-    /// Delta from mean distance
-    pub delta_value: f32,
-}
-
-/// Helper function to dequantize INT8 distance back to f32
-fn dequantize_distance(quantized: u8) -> f32 {
-    // Simple linear dequantization from [0, 255] to [0.0, 2.0] range
-    // This assumes cosine similarity distance in range [0, 2]
-    (quantized as f32) * 2.0 / 255.0
-}
-
-impl VectorCentroidMatrix {
-    /// Get distance from vector to centroid
-    pub fn get_distance(&self, vector_idx: usize, centroid_idx: usize) -> Result<f32> {
-        match self.storage_strategy {
-            VectorCentroidStorageStrategy::Full => {
-                // Direct lookup in full matrix
-                let linear_idx = vector_idx * self.num_centroids as usize + centroid_idx;
-                let byte_offset = linear_idx * 2;
-
-                if byte_offset + 2 > self.compressed_data.len() {
-                    return Err(anyhow::anyhow!("Index out of bounds"));
-                }
-
-                let quantized = u16::from_le_bytes([
-                    self.compressed_data[byte_offset],
-                    self.compressed_data[byte_offset + 1],
-                ]);
-
-                let range = self.compression_metadata.global_max_distance
-                    - self.compression_metadata.global_min_distance;
-                let scale_factor = if range > 0.0 { 65535.0 / range } else { 1.0 };
-                Ok(quantized as f32 / scale_factor + self.compression_metadata.global_min_distance)
-            }
-
-            VectorCentroidStorageStrategy::Hierarchical => {
-                // Get mean + delta if exists
-                if let Some(ref hier_data) = self.hierarchical_data {
-                    let mean = hier_data.mean_distances[centroid_idx];
-
-                    // Look for specific delta
-                    for delta in &hier_data.sparse_deltas {
-                        if delta.vector_index as usize == vector_idx
-                            && delta.centroid_index as usize == centroid_idx
-                        {
-                            return Ok(mean + delta.delta_value);
-                        }
-                    }
-
-                    // No delta found, return mean
-                    Ok(mean)
-                } else {
-                    Err(anyhow::anyhow!("Hierarchical data not available"))
-                }
-            }
-
-            VectorCentroidStorageStrategy::Sparse => {
-                // Search in sparse entries
-                if let Some(ref sparse_data) = self.sparse_data {
-                    for entry in &sparse_data.entries {
-                        if entry.vector_idx as usize == vector_idx
-                            && entry.centroid_idx as usize == centroid_idx
-                        {
-                            return Ok(dequantize_distance(entry.quantized_distance));
-                        }
-                    }
-
-                    // Not in top-k, return infinity (or max distance)
-                    Ok(f32::INFINITY)
-                } else {
-                    Err(anyhow::anyhow!("Sparse data not available"))
-                }
-            }
-        }
-    }
-}
-
-/// P×K Vector-to-centroid distance matrix reference (stored per rowgroup)
-/// Points to compressed matrix data in file, loaded on-demand during search
-///
-/// DESIGN RATIONALE:
-/// - Each rowgroup has P vectors, needs distances to all K centroids
-/// - Matrix size: P×K×4 bytes (4MB for p=1000, k=1000)
-/// - Too large to keep all in memory → on-demand loading
-/// - Critical for d₁, d₄, d₅ components in 5-component boosting
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct VectorCentroidMatrixRef {
-    /// Rowgroup this matrix belongs to
-    pub rowgroup_id: u16,
-
-    /// Number of vectors in this rowgroup (P)
-    pub num_vectors: u32,
-
-    /// Number of centroids (K, same across all rowgroups)
-    pub num_centroids: u32,
-
-    /// File offset where compressed matrix data starts
-    pub file_offset: u64,
-
-    /// Compressed size in bytes
-    pub compressed_size: u32,
-
-    /// Uncompressed size in bytes (P×K×4)
-    pub uncompressed_size: u32,
-
-    /// Compression algorithm used
-    pub compression_algorithm: String,
-
-    /// Proxima encoding metadata for efficient decompression
-    pub encoding_metadata: VectorCentroidCompressionMetadata,
-}
-
-/// Compression metadata for vector-centroid matrices
-/// Uses sophisticated encoding since distances have different characteristics per centroid
-#[derive(Debug, Clone, Default, serde::Serialize, serde::Deserialize)]
-pub struct VectorCentroidCompressionMetadata {
-    /// Per-centroid statistics for adaptive encoding
-    /// Each centroid column may have different distance distribution
-    pub centroid_stats: Vec<CentroidDistanceStats>,
-
-    /// Global normalization factors
-    pub global_min_distance: f32,
-    pub global_max_distance: f32,
-    pub global_mean_distance: f32,
-
-    /// Per-centroid encoding schemes (adaptive based on distribution)
-    pub centroid_encodings: Vec<ProximaScheme>,
-}
-
-/// Per-centroid distance statistics for adaptive compression
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct CentroidDistanceStats {
-    /// Centroid ID
-    pub centroid_id: u16,
-
-    /// Distance statistics for this centroid column
-    pub min_distance: f32,
-    pub max_distance: f32,
-    pub mean_distance: f32,
-    pub std_deviation: f32,
-
-    /// Quantization parameters
-    pub quantization_scale: f32,
-    pub quantization_offset: f32,
-}
-
-/// Proxima encoding metadata for a dimension
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ProximaMetadata {
-    /// Min value in this dimension (for delta encoding)
-    pub min_value: f32,
-
-    /// Max value in this dimension
-    pub max_value: f32,
-
-    /// Encoding scheme used
-    pub encoding: ProximaScheme,
-
-    /// Compressed size in bytes
-    pub compressed_size: u32,
-}
-
-// ProximaScheme now imported from common::proximaencoder module for code reuse
-
-/// Sparse entry in P×K matrix for boundary vectors only
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SparseEntry {
-    /// Vector index within rowgroup
-    pub vector_idx: u32,
-
-    /// Centroid index  
-    pub centroid_idx: u32,
-
-    /// Quantized distance (INT8)
-    pub quantized_distance: u8,
-}
-
-/// Sparse storage for P×K matrix with boundary detection
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct SparseData {
-    /// Number of top-k centroids stored per boundary vector
-    pub top_k: u32,
-
-    /// Sparse entries (only boundary vectors)
-    pub entries: Vec<SparseEntry>,
-
-    /// BloomFilter for fast boundary vector lookup (2KB, 1% false positive)
-    pub boundary_bloom_filter: Option<Vec<u8>>,
-
-    /// Actual sparsity ratio achieved (0.1 to 1.0)
-    pub sparsity_ratio: f32,
-}
-
-// ====== P² Matrix for Intra-Rowgroup Navigation ======
-
-/// P² Matrix: Pre-computed distances between all vectors in a rowgroup
-/// Rowgroup range for a centroid (ensures spatial locality)
-#[derive(Debug, Clone)]
-pub struct RowGroupRange {
-    /// Starting rowgroup ID for this centroid (inclusive)
-    pub start_rowgroup: u16,
-    /// Ending rowgroup ID for this centroid (inclusive)
-    pub end_rowgroup: u16,
-    /// Number of vectors in this centroid across all its rowgroups
-    pub total_vectors: u32,
-}
-
-/// Pre-computed P×(P-1)/2 distances for exact intra-rowgroup search
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct P2Matrix {
-    /// Number of vectors in this rowgroup
-    pub num_vectors: u32,
-
-    /// Upper triangle distances (P×(P-1)/2 values)
-    /// Stored as linear array, indexed by: idx = i×(2n-i-1)/2 + j - i - 1
-    pub distances: Vec<u8>, // Quantized INT8 or Binary
-
-    /// Quantization parameters for distance reconstruction
-    pub min_distance: f32,
-    pub max_distance: f32,
-
-    /// Compression strategy used
-    pub compression: ProximaScheme,
-
-    /// Size after compression
-    pub compressed_size: u32,
-}
-
-impl P2Matrix {
-    /// Get distance between vectors i and j (handles upper triangle indexing)
-    /// Delegates dequantization to StorageQuantizationEngine for consistency
-    pub fn get_distance(&self, i: usize, j: usize) -> f32 {
-        if i == j {
-            return 0.0;
-        }
-
-        // Ensure i < j for upper triangle
-        let (i, j) = if i < j { (i, j) } else { (j, i) };
-
-        // Calculate index in linear array
-        let n = self.num_vectors as usize;
-        let idx = i * (2 * n - i - 1) / 2 + j - i - 1;
-
-        // Simple dequantization from u8 to f32
-        let scale = (self.max_distance - self.min_distance) / 255.0;
-        self.distances[idx] as f32 * scale + self.min_distance
-    }
-
-    /// Get all distances for a specific vector
-    pub fn get_vector_distances(&self, vector_idx: usize) -> Vec<f32> {
-        let n = self.num_vectors as usize;
-        let mut distances = vec![0.0; n];
-
-        for (j, dist) in distances.iter_mut().enumerate() {
-            if j != vector_idx {
-                *dist = self.get_distance(vector_idx, j);
-            }
-        }
-
-        distances
-    }
-}
-
-// ====== Boundary Detection and Self-Correction Structures ======
-
-/// Spillover information between clusters
-#[derive(Debug, Clone)]
-pub struct SpilloverInfo {
-    pub from_cluster: u32,
-    pub to_cluster: u32,
-    pub spillover_ratio: f32,
-    pub vector_count: usize,
-}
-
-/// Confidence assessment for search results
-#[derive(Debug, Clone)]
-pub struct ConfidenceAssessment {
-    pub overall: f32,
-    pub signals: ConfidenceSignals,
-    pub needs_correction: bool,
-}
-
-/// Individual confidence signals
-#[derive(Debug, Clone)]
-pub struct ConfidenceSignals {
-    pub distance_uniformity: f32,
-    pub cluster_diversity: f32,
-    pub result_continuity: f32,
-    pub boundary_clarity: f32,
-}
-
-/// Correction strategy for self-correction
-#[derive(Debug, Clone)]
-pub enum CorrectionStrategy {
-    GapFilling,
-    Diversification,
-    BoundaryExploration,
-}
-
-/// Boosting strategy configuration
-#[derive(Debug, Clone)]
-pub struct BoostingStrategy {
-    pub spillover_strength: f32, // Default: 2.0
-    pub ranking_strength: f32,   // Default: 0.1
-}
-
-// ====== Enhanced Vector ID BloomFilter for Existing RowGroups ======
-
-/// Enhanced helper functions for existing RowGroupBloomFilter system
-/// Uses the existing per-rowgroup bloom_filter field instead of duplicate global registry
-impl RowGroupBloomFilter {
-    /// Fast batch lookup across multiple rowgroups using existing BloomFilters
-    /// Returns list of rowgroup IDs that might contain each vector ID
-    ///
-    /// This leverages the existing bloom_filter field in each RowGroup instead of a global registry
-    pub fn find_candidate_rowgroups_in_footer(
-        footer: &RaptorFooter,
-        vector_ids: &[String],
-    ) -> Vec<Vec<u16>> {
-        let mut results = Vec::with_capacity(vector_ids.len());
-
-        for _vector_id in vector_ids {
-            let mut candidates = Vec::new();
-
-            // Check each rowgroup's bloom filter offset (actual bloom filter needs to be loaded)
-            for rowgroup in &footer.file_metadata.row_groups {
-                if rowgroup.bloom_filter_offset.is_some() {
-                    // Deferred: Load bloom filter from file using offset and check
-                    // For now, add all rowgroups with bloom filters as candidates
-                    candidates.push(rowgroup.id);
-                }
-            }
-
-            candidates.sort_unstable();
-            results.push(candidates);
-        }
-
-        results
-    }
-
-    /// Enhanced hardware-optimized batch processing using existing RowGroup BloomFilters
-    /// Automatically detects and uses best SIMD instructions available
-    pub fn find_candidates_batch_optimized(
-        footer: &RaptorFooter,
-        vector_ids: &[String],
-    ) -> Vec<Vec<u16>> {
-        use proximadb_hardware_caps::get_hardware_capabilities;
-
-        let hw = get_hardware_capabilities();
-        let batch_size = if hw.has_avx512() {
-            16
-        } else if hw.cpu.features.avx2_support {
-            8
-        } else {
-            4
-        };
-
-        let mut results = Vec::with_capacity(vector_ids.len());
-
-        // Process in hardware-optimized batches
-        for chunk in vector_ids.chunks(batch_size) {
-            for _id in chunk {
-                let mut candidates = Vec::new();
-
-                for rowgroup in &footer.file_metadata.row_groups {
-                    if rowgroup.bloom_filter_offset.is_some() {
-                        // Deferred: Load bloom filter from file using offset and check
-                        // For now, add all rowgroups with bloom filters as candidates
-                        candidates.push(rowgroup.id);
-                    }
-                }
-
-                candidates.sort_unstable();
-                results.push(candidates);
-            }
-        }
-
-        results
-    }
-}
-
-/// K² matrix for inter-centroid navigation
-/// Stores pairwise distances between centroids (upper triangle only)
-#[derive(Debug, Clone)]
-pub struct K2Matrix {
-    /// Number of centroids
-    pub num_centroids: u32,
-
-    /// Upper triangle distances (K×(K-1)/2 values)
-    /// Stored as linear array, indexed by: idx = i×(2n-i-1)/2 + j - i - 1
-    pub distances: Vec<u8>, // Quantized INT8 or INT16
-
-    /// Quantization parameters for distance reconstruction
-    pub min_distance: f32,
-    pub max_distance: f32,
-
-    /// Compression strategy used
-    pub compression: ProximaScheme,
-
-    /// Size after compression
-    pub compressed_size: u32,
-}
-
-impl K2Matrix {
-    /// Get distance between centroids i and j (handles upper triangle indexing)
-    /// Delegates dequantization to StorageQuantizationEngine for consistency
-    pub fn get_distance(&self, i: usize, j: usize) -> f32 {
-        if i == j {
-            return 0.0;
-        }
-
-        // Ensure i < j for upper triangle
-        let (i, j) = if i < j { (i, j) } else { (j, i) };
-
-        let n = self.num_centroids as usize;
-        let idx = i * (2 * n - i - 1) / 2 + j - i - 1;
-
-        if idx < self.distances.len() {
-            // Simple dequantization from u8 to f32
-            let scale = (self.max_distance - self.min_distance) / 255.0;
-            self.distances[idx] as f32 * scale + self.min_distance
-        } else {
-            self.max_distance // Fallback
-        }
-    }
-}
-
-// ====== Runtime Search Result Structures (NOT STORED) ======
-
-/// Runtime result from Phase 1 boundary detection
-/// This is computed during search using K² matrix - NOT stored
-#[derive(Debug, Clone)]
-pub struct BoundaryDetectionResult {
-    /// Primary centroids selected by distance ranking
-    pub primary_centroids: Vec<u16>,
-
-    /// Neighbor centroids discovered via boundary rule
-    /// Each tuple: (from_centroid, to_centroid, boundary_ratio)
-    pub boundary_expansions: Vec<(u16, u16, f32)>,
-
-    /// Final expanded set of centroids to search
-    pub expanded_centroids: Vec<u16>,
-
-    /// Statistics for debugging/monitoring
-    pub boundary_ratio_threshold: f32, // Threshold used (default 0.8)
-    pub expansion_count: usize, // Number of centroids added
-}
-
-/// Runtime result from Phase 2 spillover detection  
-/// This is computed during search using P×K matrix - NOT stored
-#[derive(Debug, Clone)]
-pub struct SpilloverDetectionResult {
-    /// Map of centroid_id -> list of spillover target centroids
-    pub spillover_map: HashMap<u16, Vec<u16>>,
-
-    /// Spillover percentages for each centroid
-    pub spillover_percentages: HashMap<u16, f32>,
-
-    /// Centroids added via recursive spillover detection
-    pub recursive_expansions: Vec<u16>,
-
-    /// Final complete set of centroids after spillover
-    pub final_centroids: Vec<u16>,
-
-    /// Statistics for debugging/monitoring
-    pub spillover_threshold: f32, // Threshold used (default 0.15)
-    pub max_spillover_percentage: f32,  // Highest spillover % found
-    pub total_spillover_vectors: usize, // Total vectors that spilled
-}
-
-/// Configuration for boundary detection (runtime)
-#[derive(Debug, Clone)]
-pub struct BoundaryDetectionConfig {
-    /// Boundary ratio threshold (default 0.8)
-    /// If d(Q,C_i)/d(Q,C_j) > threshold, consider C_j as boundary neighbor
-    pub boundary_ratio_threshold: f32,
-
-    /// Maximum expansion budget as percentage (default 0.2 = 20%)
-    pub expansion_budget_percent: f32,
-
-    /// Minimum distance for boundary detection (default: mean - 2σ)
-    pub min_boundary_distance: f32,
-
-    /// Maximum neighbor candidates to consider (default: K/4)
-    pub max_neighbor_candidates: usize,
-}
-
-impl Default for BoundaryDetectionConfig {
-    fn default() -> Self {
-        Self {
-            boundary_ratio_threshold: 0.8,
-            expansion_budget_percent: 0.2,
-            min_boundary_distance: 0.0,  // Will be computed from stats
-            max_neighbor_candidates: 25, // For K=100
-        }
-    }
-}
-
-/// Configuration for spillover detection (runtime)
-#[derive(Debug, Clone)]
-pub struct SpilloverDetectionConfig {
-    /// Spillover percentage threshold (default 0.15 = 15%)
-    /// If >threshold% of vectors spill to another centroid, expand search
-    pub spillover_threshold: f32,
-
-    /// Distance ratio for spillover detection (default 0.9)
-    /// Vector spills if d(V,C_other) < d(V,C_own) × ratio
-    pub distance_ratio_threshold: f32,
-
-    /// Maximum recursive depth for cascading spillovers (default 2)
-    pub max_recursive_depth: usize,
-
-    /// Maximum spillover centroids to add (default K/2)
-    pub max_spillover_centroids: usize,
-}
-
-impl Default for SpilloverDetectionConfig {
-    fn default() -> Self {
-        Self {
-            spillover_threshold: 0.15,
-            distance_ratio_threshold: 0.9,
-            max_recursive_depth: 2,
-            max_spillover_centroids: 50, // For K=100
-        }
-    }
-}
+//! Re-export shim — the RAPTOR common types have been hoisted to the
+//! `proximadb-raptor-common` crate as the raptor cycle-hub inversion (unblocks the
+//! raptor engine extraction). All `crate::storage::engines::raptor::common::*` paths
+//! (the 2 direct importers + the `Predicate` re-export in `mod.rs`, used widely)
+//! resolve unchanged through this glob.
+//!
+//! See `docs/12-design/ROOT_CRATE_DECOMPOSITION_ENGINES_EXTRACTION_2026_07_12.adoc`.
+
+pub use proximadb_raptor_common::*;
